+++ CHANGELOG.md
@@ -4,7 +4,19 @@
 
 ## [Unreleased] - ReleaseDate
 
-[Commits](https://github.com/twitch-rs/twitch_oauth2/compare/v0.11.0...Unreleased)
+[Commits](https://github.com/twitch-rs/twitch_oauth2/compare/v0.11.1...Unreleased)
+
+## [v0.11.1] - 2023-02-01
+
+[Commits](https://github.com/twitch-rs/twitch_oauth2/compare/v0.11.0...v0.11.1)
+
+### Added
+
+- Added scopes for shoutouts
+
+### Changed
+
+- Marked `channel_subscriptions` as deprecated.
 
 ## [v0.11.0] - 2023-01-24